requirements.txt
selenium
pytest
allure-pytest
flake8
webdriver-manager